+++ packages/react/src/index.ts
@@ -3,6 +3,7 @@ import {
   cleanMultiLine,
   css,
   isQuarkCss,
+  PartialPropsOfVariantsMap,
   PropsOfVariantsMap,
   QuarkConfig,
   QuarkCss,
@@ -23,21 +24,23 @@ export type QuarkComponentPolymorphicProps<
   As extends ElementType | never,
   Element extends ElementType,
   VariantsMap extends QuarkVariantsMap,
-  DefaultProps extends PartialComponentProps<Element>
+  Defaults extends PartialPropsOfVariantsMap<VariantsMap>,
+  DefaultComponentProps extends PartialComponentProps<Element>
 > = { as?: As } & Assign<
-  Assign<ComponentProps<IsNever<As, Element>>, Partial<DefaultProps>>,
-  PropsOfVariantsMap<VariantsMap>
+  Assign<ComponentProps<IsNever<As, Element>>, Partial<DefaultComponentProps>>,
+  PropsOfVariantsMap<VariantsMap, Defaults>
 >
 
 export interface QuarkComponent<
   Element extends ElementType,
   VariantsMap extends QuarkVariantsMap,
-  DefaultProps extends PartialComponentProps<Element>
+  Defaults extends PartialPropsOfVariantsMap<VariantsMap>,
+  DefaultComponentProps extends PartialComponentProps<Element>
 > {
   <As extends ElementType | never = never>(
-    props: QuarkComponentPolymorphicProps<As, Element, VariantsMap, DefaultProps>
+    props: QuarkComponentPolymorphicProps<As, Element, VariantsMap, Defaults, DefaultComponentProps>
   ): ReactElement<any, any> | null
-  CSS: QuarkCss<VariantsMap>
+  CSS: QuarkCss<VariantsMap, Defaults>
 }
 
 type IsNever<T, A> = [T] extends [never] ? A : T
@@ -45,19 +48,29 @@ type IsNever<T, A> = [T] extends [never] ? A : T
 // type Assign<A, B> = Omit<A, keyof B> & B
 type Assign<A, B> = Omit<A, keyof B | 'as'> & B
 
-export type QuarkComponentVariantsMap<C> = C extends QuarkComponent<any, infer V, any> ? V : never
-export type QuarkComponentVariants<C> = PropsOfVariantsMap<QuarkComponentVariantsMap<C>>
-export type PropsOf<C extends ElementType> = C extends QuarkComponent<infer E, infer V, infer D>
-  ? QuarkComponentPolymorphicProps<never, E, V, D>
-  : ComponentProps<C>
+export type QuarkComponentVariantProps<C> = C extends QuarkComponent<any, infer V, infer D, any>
+  ? PropsOfVariantsMap<V, D>
+  : never
+
+export type PropsOf<COMP extends ElementType> = COMP extends QuarkComponent<
+  infer E,
+  infer C,
+  infer D,
+  infer CD
+>
+  ? QuarkComponentPolymorphicProps<never, E, C, D, CD>
+  : ComponentProps<COMP>
 
 export type PartialComponentProps<Element extends ElementType> = Partial<ComponentProps<Element>>
 
-type AnyQuarkCss = QuarkCss<any>
+type AnyQuarkCss = QuarkCss<any, any>
 
-type MaybeQuarkConfig<VariantsMap extends QuarkVariantsMap> =
-  | QuarkConfig<VariantsMap>
-  | QuarkCss<VariantsMap>
+type MaybeQuarkConfig<
+  VariantsMap extends QuarkVariantsMap,
+  Defaults extends PartialPropsOfVariantsMap<VariantsMap>
+> =
+  | QuarkConfig<VariantsMap, Defaults>
+  | QuarkCss<VariantsMap, Defaults>
   | string[]
   //Hack so that typescript can narrow type errors to QuarkConfig instead of the whole parameter
   | (string & { quark?: VariantsMap })
@@ -65,31 +78,37 @@ type MaybeQuarkConfig<VariantsMap extends QuarkVariantsMap> =
 export type StyledFn = <
   Element extends ElementType,
   VariantsMap extends QuarkVariantsMap = {},
+  Defaults extends PartialPropsOfVariantsMap<VariantsMap> = {},
   DefaultProps extends PartialComponentProps<Element> = {}
 >(
   element: Element,
-  configOrCssOrClassStrings: MaybeQuarkConfig<VariantsMap>,
+  configOrCssOrClassStrings: MaybeQuarkConfig<VariantsMap, Defaults>,
   defaultComponentProps?: DefaultProps
-) => QuarkComponent<Element, VariantsMap, DefaultProps>
+) => QuarkComponent<Element, VariantsMap, Defaults, DefaultProps>
 
 export type Styled = StyledFn & {
   [K in keyof JSX.IntrinsicElements]: {
-    <VariantsMap extends QuarkVariantsMap = {}, DefaultProps extends PartialComponentProps<K> = {}>(
-      configOrCssOrClassStrings: MaybeQuarkConfig<VariantsMap>,
+    <
+      VariantsMap extends QuarkVariantsMap = {},
+      Defaults extends PartialPropsOfVariantsMap<VariantsMap> = {},
+      DefaultProps extends PartialComponentProps<K> = {}
+    >(
+      configOrCssOrClassStrings: MaybeQuarkConfig<VariantsMap, Defaults>,
       defaultComponentProps?: DefaultProps
-    ): QuarkComponent<K, VariantsMap, DefaultProps>
+    ): QuarkComponent<K, VariantsMap, Defaults, DefaultProps>
   }
 }
 
 function _styled<
   Element extends ElementType,
-  VariantsMap extends QuarkVariantsMap = {},
+  VariantsMap extends QuarkVariantsMap,
+  Defaults extends PartialPropsOfVariantsMap<VariantsMap>,
   DefaultProps extends PartialComponentProps<Element> = {}
 >(
   element: Element,
-  configOrCssOrClassStrings: MaybeQuarkConfig<VariantsMap>,
+  configOrCssOrClassStrings: MaybeQuarkConfig<VariantsMap, Defaults>,
   defaultComponentProps?: DefaultProps
-): QuarkComponent<Element, VariantsMap, DefaultProps> {
+): QuarkComponent<Element, VariantsMap, Defaults, DefaultProps> {
   const quark = isQuarkCss(configOrCssOrClassStrings)
     ? (configOrCssOrClassStrings as AnyQuarkCss)
     : !isStrings(configOrCssOrClassStrings)
@@ -110,9 +129,7 @@ function _styled<
     }
   } else {
     Component = ({ children, className: _className, as, ...props }, ref) => {
-      const baseClass = arrayify(configOrCssOrClassStrings as string | string[])
-        .map(cleanMultiLine)
-        .join(' ')
+      const baseClass = cleanMultiLine(arrayify(configOrCssOrClassStrings as string | string[]).join(' '))
 
       const className = _className ? `${_className} ${baseClass}` : baseClass
 
@@ -128,8 +145,6 @@ const isStrings = (value: any): value is string | string[] => {
   return typeof value === 'string' || Array.isArray(value)
 }
 
-//Re-export @quark/core
-
 export const styled: Styled = new Proxy(_styled, {
   get(target, prop) {
     if (typeof prop !== 'string') throw new Error(`styled: invalid prop \`${prop.toString()}\` `)
@@ -139,5 +154,6 @@ export const styled: Styled = new Proxy(_styled, {
   },
 }) as any
 
+//Re-export @quark/core
 export { css, isQuarkCss }
 export type { QuarkConfig, QuarkCss, QuarkVariantsMap }