+++ packages/react/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @quarkcss/react
 
+## 0.5.2
+
+### Patch Changes
+
+- Variant Props are now required unless default or null value is set
+- Updated dependencies
+  - @quarkcss/core@0.5.2
+
 ## 0.5.1
 
 ### Patch Changes